+++ frontend/src/customHooks/useDebounce.tsx
@@ -1,0 +1,21 @@
+import { useState, useEffect } from 'react'
+
+function useDebounce(value: any, delay: number) {
+
+    const [debouncedValue, setDebouncedValue] = useState(value)
+    useEffect(() => {
+
+        const handler = setTimeout(() => {
+            setDebouncedValue(value)
+        }, delay)
+
+        return () => {
+            clearTimeout(handler)
+        }
+
+    }, [value, delay])
+
+    return debouncedValue
+}
+
+export default useDebounce
